refactor (location) remove used location functions from data generator
