+++ src/data-generator/location.ts
@@ -54,26 +54,6 @@ export async function getLocations() {
   )
 }
 
-export async function getLocationById(fhirId: string) {
-  const url = `${GATEWAY_HOST}/location/${fhirId}`
-
-  const res = await fetch(url, {
-    method: 'GET'
-  })
-
-  const response = await res.json()
-  return response
-}
-
-export async function getOffices() {
-  const resCRVSOffices = await fetch(
-    `${GATEWAY_HOST}/location?type=CRVS_OFFICE&_count=0&status=active`
-  )
-
-  const officeList = await resCRVSOffices.json()
-  return officeList?.entry
-}
-
 export async function getFacilities() {
   const resCRVSOffices = await fetch(
     `${GATEWAY_HOST}/location?type=CRVS_OFFICE&_count=0`
